+++ flutter_frontend/packages/libmsgr/lib/src/database/database.dart
@@ -4,6 +4,7 @@ import 'package:libmsgr/src/database/migrations/000_initial_migration.dart';
 import 'package:libmsgr/src/database/migrations/001_add_messages_and_contacts.dart';
 import 'package:libmsgr/src/database/migrations/002_profile_preferences.dart';
 import 'package:libmsgr/src/database/migrations/003_outgoing_message_queue.dart';
+import 'package:libmsgr/src/database/migrations/004_message_delivery_status.dart';
 import 'package:logging/logging.dart';
 import 'package:sqflite_common/src/sql_builder.dart';
 import 'package:sqflite_sqlcipher/sqflite.dart';
@@ -18,6 +19,7 @@ const List<Migration<DatabaseMigrationData>> migrations = [
   Migration(3, upgradeFromV2ToV3),
   Migration(4, upgradeFromV3ToV4),
   Migration(5, upgradeFromV4ToV5),
+  Migration(6, upgradeFromV5ToV6),
 ];
 
 class DatabaseService {